testproj/requirements.txt
Django==1.6.2
django-tastypie
djangorestframework
